+++ app/schemas/movie.py
@@ -10,7 +10,7 @@ class MovieCreate(SQLModel):
     release_date: date | None = None
     genre_id: int | None = None
     poster_url: str | None = None
-    rating: int | None = None
+    rating: float | None = None
     is_published: bool | None = None
     file_key: str | None = None  # Delete the None in production
 
@@ -25,6 +25,6 @@ class MovieUpdate(SQLModel):
     release_date: date | None = None
     genre_id: int | None = None
     poster_url: str | None = None
-    rating: int | None = None
+    rating: float | None = None
     is_published: bool | None = None
     file_key: str | None = None # Delete the None in production
